requirements.txt
mcp
python-dotenv
pandas
selenium
bs4
fastapi 
python-multipart
uvicorn 
google-api-python-client
matplotlib
pyppeteer
elasticsearch
boto3
kafka-python

pymupdf
pillow
rapidocr_onnxruntime
pydantic
opencv-python
testresources
onnx
onnxruntime
transformers
torch
scikit-learn
pytesseract
openai